Dev: Polish the sbd feature.
Move find_candidate_sbd to utils.py
Import glob from file beginning
Use `sbd -d` instead of hexdump
Use crmsh.utils.parse_sysconfig instead of conf_parser
Add doc
