+++ preflight_check/config.py
@@ -6,4 +6,4 @@ REMOVE_PORT = "firewall-cmd --zone=public --remove-port={port}/udp"
 ADD_PORT = "firewall-cmd --zone=public --add-port={port}/udp"
 FENCE_HISTORY = "stonith_admin -h {node}"
 SBD_CONF = "/etc/sysconfig/sbd"
-SBD_CHECK_CMD = "hexdump -C {dev} -n 7|head -1 |cut -d '|' -f 2"
+SBD_CHECK_CMD = "sbd -d {dev} dump"
